feat: Use external import map plugin
This make use of the external [rollup-plugin-import-map](https://github.com/trygve-lie/rollup-plugin-import-map) plugin to deal with applying import maps to the code leaving this module only dealing with reading Eik config and fetching import maps from an Eik server.

In this process I would like to suggest changing a couple of things which are breaking:

Currently this plugin can take an import map as an object together with ex an Array of import maps to load from an Eik server. This is done as follow:

```js
plugin({
    imports: {
            'lit-html/lit-html': 'https://cdn.eik.dev/lit-html/v2'
    },
    urls: ['https://cdn.eik.dev/......./import-map.json'],
})
```

The problem with this is that the import map passed on is not a full import map. Its a subset of an import map. If it should be possible to pass in a full import map it should be like this:

```js
plugin({
    imports: {
        imports: {
                'lit-html/lit-html': 'https://cdn.eik.dev/lit-html/v2'
        }
    },
    urls: ['https://cdn.eik.dev/......./import-map.json'],
})
```

This is a bit weird. My suggestion is that we change to do the following:

 * The `imports` argument is renamed to `maps`.
 * The `maps` argument can take an array of import maps.

Then it will be as follow:

```js
plugin({
    maps: [{
        imports: {
                'lit-html/lit-html': 'https://cdn.eik.dev/lit-html/v2'
        }
    }],
    urls: ['https://cdn.eik.dev/......./import-map.json'],
})
```

Currently the argument for pointing to the eik.json file is named `path`. I would like to rename this to `config`.

BREAKING CHANGE: Restructure behaviour of and naming of arguments